feat(home): sửa lại ui màn home - giao diện cũ cứ bẩn bẩn sao á :)) cái mới chưa ok nhưng ổn hơn khá nhiều
