+++ app/src/main/java/com/pwhs/quickmem/presentation/StandardScaffold.kt
@@ -5,23 +5,14 @@ import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.FabPosition
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme.colorScheme
 import androidx.compose.material3.MaterialTheme.typography
@@ -33,7 +24,6 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -43,25 +33,19 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import androidx.navigation.compose.currentBackStackEntryAsState
-import com.airbnb.lottie.compose.LottieAnimation
 import com.airbnb.lottie.compose.LottieCompositionSpec
 import com.airbnb.lottie.compose.LottieConstants
 import com.airbnb.lottie.compose.animateLottieCompositionAsState
 import com.airbnb.lottie.compose.rememberLottieComposition
 import com.pwhs.quickmem.R
-import com.pwhs.quickmem.presentation.app.home.components.StreakCalendar
 import com.pwhs.quickmem.presentation.component.BottomSheetItem
-import com.pwhs.quickmem.ui.theme.streakTextColor
-import com.pwhs.quickmem.ui.theme.streakTitleColor
 import com.ramcosta.composedestinations.generated.destinations.CreateClassScreenDestination
 import com.ramcosta.composedestinations.generated.destinations.CreateFolderScreenDestination
 import com.ramcosta.composedestinations.generated.destinations.CreateStudySetScreenDestination
@@ -81,23 +65,12 @@ fun StandardScaffold(
         BottomNavItem.Profile
     ),
     content: @Composable (PaddingValues) -> Unit,
-    viewModel: StandardViewModel = hiltViewModel(),
 ) {
     val sheetSelectCreateState = rememberModalBottomSheetState()
     var showBottomSheetCreate by remember {
         mutableStateOf(false)
     }
     val navigator = navController.rememberDestinationsNavigator()
-    val streakBottomSheet = rememberModalBottomSheetState()
-    var showStreakBottomSheet by remember {
-        mutableStateOf(false)
-    }
-    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(R.raw.lottie_fire_streak))
-    val progress by animateLottieCompositionAsState(
-        composition = composition,
-        iterations = LottieConstants.IterateForever,
-    )
-    val uiState by viewModel.uiState.collectAsState()
 
     Scaffold(
         bottomBar = {
@@ -205,49 +178,6 @@ fun StandardScaffold(
                 }
             }
         },
-        floatingActionButton = {
-            if (showBottomBar) {
-                Card(
-                    onClick = {
-                        showStreakBottomSheet = true
-                    },
-                    shape = CircleShape,
-                    elevation = CardDefaults.elevatedCardElevation(
-                        defaultElevation = 8.dp
-                    ),
-                    colors = CardDefaults.cardColors(
-                        containerColor = colorScheme.surface
-                    ),
-                    border = BorderStroke(
-                        width = 2.dp,
-                        color = Color.White
-                    )
-                ) {
-                    Row(
-                        horizontalArrangement = Arrangement.spacedBy(8.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                        modifier = Modifier
-                            .padding(5.dp)
-                            .padding(horizontal = 8.dp)
-                    ) {
-                        Image(
-                            painter = painterResource(id = R.drawable.ic_fire),
-                            contentDescription = stringResource(R.string.txt_streak),
-                            contentScale = ContentScale.Crop,
-                            modifier = Modifier.size(30.dp)
-                        )
-                        Text(
-                            text = "${uiState.streakCount}",
-                            style = typography.titleLarge.copy(
-                                color = colorScheme.primary,
-                                fontWeight = FontWeight.Bold,
-                            )
-                        )
-                    }
-                }
-            }
-        },
-        floatingActionButtonPosition = FabPosition.Start,
     ) { innerPadding ->
         content(innerPadding)
         if (showBottomSheetCreate) {
@@ -285,51 +215,6 @@ fun StandardScaffold(
                             showBottomSheetCreate = false
                             navController.navigate(CreateClassScreenDestination.route)
                         }
-                    )
-                }
-            }
-        }
-        if (showStreakBottomSheet) {
-            ModalBottomSheet(
-                onDismissRequest = {
-                    showStreakBottomSheet = false
-                },
-                sheetState = streakBottomSheet,
-            ) {
-                Column(
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    LottieAnimation(
-                        composition = composition,
-                        progress = { progress },
-                        modifier = Modifier
-                            .width(100.dp)
-                            .height(100.dp)
-                    )
-                    Text(
-                        text = "${uiState.streakCount}",
-                        style = typography.titleLarge.copy(
-                            color = streakTitleColor,
-                            fontWeight = FontWeight.Bold,
-                            fontSize = 52.sp
-                        )
-                    )
-                    Text(
-                        text = when (uiState.streakCount) {
-                            1 -> stringResource(R.string.txt_day_streak)
-                            else -> stringResource(R.string.txt_days_streak)
-                        },
-                        style = typography.titleLarge.copy(
-                            color = streakTextColor,
-                            fontWeight = FontWeight.Bold
-                        ),
-                        modifier = Modifier.padding(bottom = 16.dp)
-                    )
-                    Text(
-                        text = stringResource(R.string.txt_practice_every_day),
-                    )
-                    StreakCalendar(
-                        streakDates = uiState.streakDates
                     )
                 }
             }
