--- app/src/main/java/com/pwhs/quickmem/presentation/StandardUiAction.kt
@@ -1,5 +1,0 @@
-package com.pwhs.quickmem.presentation
-
-sealed class StandardUiAction {
-    data object UpdateStreak : StandardUiAction()
-}
