--- app/src/main/java/com/pwhs/quickmem/presentation/StandardUiState.kt
@@ -1,11 +1,0 @@
-package com.pwhs.quickmem.presentation
-
-import com.pwhs.quickmem.domain.model.streak.StreakModel
-import java.time.LocalDate
-
-data class StandardUiState(
-    val streakCount: Int = 0,
-    val streaks: List<StreakModel> = emptyList(),
-    val streakDates: List<LocalDate> = emptyList(),
-    val isLoading: Boolean = false,
-)
