+++ app/src/main/java/com/pwhs/quickmem/presentation/app/study_set/studies/flip/FlipFlashCardScreen.kt
@@ -44,8 +44,8 @@ import com.msusman.compose.cardstack.rememberStackState
 import com.pwhs.quickmem.R
 import com.pwhs.quickmem.core.data.enums.LearnFrom
 import com.pwhs.quickmem.domain.model.flashcard.FlashCardResponseModel
-import com.pwhs.quickmem.presentation.StandardUiAction
-import com.pwhs.quickmem.presentation.StandardViewModel
+import com.pwhs.quickmem.presentation.app.home.HomeUiAction
+import com.pwhs.quickmem.presentation.app.home.HomeViewModel
 import com.pwhs.quickmem.presentation.app.study_set.component.StudyCardBottomSheet
 import com.pwhs.quickmem.presentation.app.study_set.studies.flip.component.FlipFlashCardButton
 import com.pwhs.quickmem.presentation.app.study_set.studies.flip.component.FlipFlashCardFinish
@@ -70,7 +70,7 @@ fun FlipFlashCardScreen(
     modifier: Modifier = Modifier,
     resultNavigator: ResultBackNavigator<Boolean>,
     viewModel: FlipFlashCardViewModel = hiltViewModel(),
-    standardViewModel: StandardViewModel = hiltViewModel(),
+    homeViewModel: HomeViewModel = hiltViewModel()
 ) {
     val uiState by viewModel.uiState.collectAsState()
     val context = LocalContext.current
@@ -79,7 +79,7 @@ fun FlipFlashCardScreen(
         viewModel.uiEvent.collect { event ->
             when (event) {
                 FlipFlashCardUiEvent.Finished -> {
-                    standardViewModel.onEvent(StandardUiAction.UpdateStreak)
+                    homeViewModel.onEvent(HomeUiAction.UpdateStreak)
                     Toast.makeText(
                         context,
                         context.getString(R.string.txt_you_have_finished),
