+++ app/src/main/java/com/pwhs/quickmem/ui/theme/Color.kt
@@ -6,7 +6,7 @@ val premiumColor = Color(0xFFfa6c3e)
 val incorrectColor = Color(0xFF860010)
 val correctColor = Color(0xFF6c9184)
 val streakTextColor = Color(0xFFf2ac40)
-val streakTitleColor = Color(0xFFf2ac40)
+val streakTitleColor = Color(0xFFFC9502)
 val borderTopStreakTextColor = Color(0xFFf2ac40)
 
 val primaryLight = Color(0xFF2BB3CD)
